refactor dirty dozen into FSD
