+++ src/fsd/0-app/routing/desktop-routes.tsx
@@ -1,11 +1,11 @@
 ﻿import React from 'react';
 import { RouteObject } from 'react-router-dom';
 
+import { dirtyDozenLazyRoute } from '@/fsd/1-pages/learn-dirty-dozen';
 import { mowLookupDesktopLazyRoute } from '@/fsd/1-pages/learn-mow';
 import { versusLazyRoute } from '@/fsd/1-pages/learn-versus';
 import { campaignProgressionLazyRoute } from '@/fsd/1-pages/plan-campaign-progression';
 import { guildApiLazyRoute } from '@/v2/pages/guild-api/guild-api.route';
-import { dirtyDozenLazyRoute } from 'src/v2/pages/dirty-dozen/dirty-dozen.route';
 import { faqLazyRoute } from 'src/v2/pages/faq/faq.route';
 import { guidesLazyRoute } from 'src/v2/pages/guides/guides.route';
 import { guildLazyRoute } from 'src/v2/pages/guild/guild.route';
