+++ src/fsd/1-pages/learn-dirty-dozen/dirty-dozen-table.css
