+++ src/fsd/1-pages/learn-dirty-dozen/dirty-dozen.menu-item.tsx
@@ -1,0 +1,6 @@
+﻿import DirtyLensIcon from '@mui/icons-material/DirtyLens';
+import React from 'react';
+
+import { MenuItemTP } from '@/fsd/5-shared/ui';
+
+export const dirtyDozenMenuItem = new MenuItemTP('Dirty Dozen', <DirtyLensIcon />, '/learn/dirtyDozen');
