+++ src/fsd/1-pages/learn-dirty-dozen/index.ts
@@ -1,0 +1,2 @@
+export { dirtyDozenLazyRoute } from './dirty-dozen.route';
+export { dirtyDozenMenuItem } from './dirty-dozen.menu-item';
