--- src/v2/pages/dirty-dozen/dirty-dozen.menu-item.tsx
@@ -1,6 +1,0 @@
-﻿import DirtyLensIcon from '@mui/icons-material/DirtyLens';
-import React from 'react';
-
-import { MenuItem } from 'src/v2/models/menu-item';
-
-export const dirtyDozenMenuItem = new MenuItem('Dirty Dozen', <DirtyLensIcon />, '/learn/dirtyDozen');
